DBZ-736 DBZ-733 Removed MySQL legacy parser and consolidate parser tests
