+++ debezium-connector-mysql/src/test/java/io/debezium/connector/mysql/MySqlConnectorRegressionIT.java
@@ -29,14 +29,10 @@ import org.apache.kafka.connect.data.Struct;
 import org.fest.assertions.Delta;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TestRule;
 
 import io.debezium.config.Configuration;
 import io.debezium.connector.mysql.MySqlConnectorConfig.SnapshotMode;
-import io.debezium.connector.mysql.junit.SkipForLegacyParser;
-import io.debezium.connector.mysql.junit.SkipTestForLegacyParser;
 import io.debezium.data.Envelope;
 import io.debezium.doc.FixFor;
 import io.debezium.embedded.AbstractConnectorTest;
@@ -58,9 +54,6 @@ public class MySqlConnectorRegressionIT extends AbstractConnectorTest {
 
     private Configuration config;
 
-    @Rule
-    public final TestRule skip = new SkipTestForLegacyParser();
-
     @Before
     public void beforeEach() {
         stopConnector();
@@ -80,7 +73,6 @@ public class MySqlConnectorRegressionIT extends AbstractConnectorTest {
 
     @Test
     @FixFor("DBZ-61")
-    @SkipForLegacyParser
     public void shouldConsumeAllEventsFromDatabaseUsingBinlogAndNoSnapshot() throws SQLException, InterruptedException {
         // Use the DB configuration to define the connector's configuration ...
         config = DATABASE.defaultConfig()
@@ -340,7 +332,6 @@ public class MySqlConnectorRegressionIT extends AbstractConnectorTest {
 
     @Test
     @FixFor("DBZ-61")
-    @SkipForLegacyParser
     public void shouldConsumeAllEventsFromDatabaseUsingBinlogAndNoSnapshotAndConnectTimesTypes() throws SQLException, InterruptedException {
         // Use the DB configuration to define the connector's configuration ...
         config = DATABASE.defaultConfig()
@@ -783,7 +774,6 @@ public class MySqlConnectorRegressionIT extends AbstractConnectorTest {
 
     @Test
     @FixFor("DBZ-147")
-    @SkipForLegacyParser
     public void shouldConsumeAllEventsFromDecimalTableInDatabaseUsingBinlogAndNoSnapshot() throws SQLException, InterruptedException {
         // Use the DB configuration to define the connector's configuration ...
         config = DATABASE.defaultConfig()
