+++ debezium-connector-mysql/src/test/java/io/debezium/connector/mysql/TopicNameSanitizationIT.java
@@ -15,13 +15,9 @@ import org.apache.kafka.connect.data.Struct;
 import org.apache.kafka.connect.source.SourceRecord;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TestRule;
 
 import io.debezium.config.Configuration;
-import io.debezium.connector.mysql.junit.SkipForLegacyParser;
-import io.debezium.connector.mysql.junit.SkipTestForLegacyParser;
 import io.debezium.data.VerifyRecord;
 import io.debezium.doc.FixFor;
 import io.debezium.embedded.AbstractConnectorTest;
@@ -32,16 +28,12 @@ import io.debezium.util.Testing;
  *
  * @author Gunnar Morling
  */
-@SkipForLegacyParser
 public class TopicNameSanitizationIT extends AbstractConnectorTest {
 
     private static final Path DB_HISTORY_PATH = Testing.Files.createTestingPath("file-db-history-topic-name-sanitization.txt")
                                                              .toAbsolutePath();
     private final UniqueDatabase DATABASE = new UniqueDatabase("topic-name-sanitization-it", "topic_name_sanitization_test")
             .withDbHistoryPath(DB_HISTORY_PATH);
-
-    @Rule
-    public final TestRule skip = new SkipTestForLegacyParser();
 
     private Configuration config;
